+++ src/Autoencoder.py
@@ -90,7 +90,7 @@ class OutputLayer:
 
     def buildTrainer(self):
         """ Trains the hidden layer. """
-        self.labelTensor = tf.placeholder(tf.float32, [2])
+        self.labelTensor = tf.placeholder(tf.int32, [2])
         print(self.labelTensor.get_shape())
         print(self.zo.get_shape())
         self.loss = tf.reduce_mean(tf.nn.sparse_softmax_cross_entropy_with_logits(labels = self.labelTensor, logits = self.zo))
